% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART2.R
\name{DART2_list}
\alias{DART2_list}
\title{DART2 function based on existed aggregation tree. This function does not construct the aggregation tree. Instead, the aggregation tree must be provided as an input parameter to the function in the form of a list of matrices.}
\usage{
DART2_list(alpha = 0.05, Llist, T1)
}
\arguments{
\item{alpha}{(numeric) Vector of desired FDR, default value is 0.05.}

\item{Llist}{(list of matrices) Aggregation tree presented in the form of matrix list.}

\item{T1}{(vector) Input test statistics}
}
\value{
(list of vectors) Index of rejected nodes on each layers
}
\description{
DART2 function based on existed aggregation tree. This function does not construct the aggregation tree. Instead, the aggregation tree must be provided as an input parameter to the function in the form of a list of matrices.
}
