% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DART2.R
\name{DART2_order}
\alias{DART2_order}
\title{DART2 function based on ordered hypotheses. The function will automatically construct the aggregation tree based on the order of hypotheses.}
\usage{
DART2_order(out, alpha = 0.05, L = 5)
}
\arguments{
\item{out}{(data.frame) Data frame with variables 1) "hypo": index of hypothesis; 2) "pval": p-values for each hypothesis; 3)"z": test statistics for each hypothesis; 4) "orderp": the order of hypothesis based on external information.}

\item{alpha}{(numeric) Vector of desired FDR, default value is 0.05.}

\item{L}{(numeric) Maximum number of layers of the aggregation tree.}
}
\value{
(data.frame) Data frame informing rejected nodes on each layer.
}
\description{
DART2 function based on ordered hypotheses. The function will automatically construct the aggregation tree based on the order of hypotheses.
}
