% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_list.R
\docType{data}
\name{Sample_list}
\alias{Sample_list}
\title{Sample data with exist aggregation tree}
\format{
A List contains two element for 1000 hypothesis: 1) Atree: aggregation tree present in the form of a list by layers; 2) stat: test statistics presented in form of z-values.
}
\usage{
Sample_list
}
\description{
Sample data with exist aggregation tree
}
\keyword{datasets}
