% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_order.R
\docType{data}
\name{Sample_order}
\alias{Sample_order}
\title{Sample data with ordered hypothesis}
\format{
A data frame contains the testing results for 1000 hypothesis. The data frame contains 4 variables 1) "hypo": index of hypothesis; 2) "pval": p-values for each hypothesis; 3)"z": test statistics for each hypothesis; 4) "orderp": the order of hypothesis based on external information.
}
\usage{
Sample_order
}
\description{
Sample data with ordered hypothesis
}
\keyword{datasets}
